Merge v5 into master

+++ se_code/copy_topics.py
@@ -11,8 +11,11 @@ def delete_topics(client):
     ids_to_delete = []
     for t in topics:
         ids_to_delete.append(t['saved_concept_id'])
-    client.delete('concepts/saved', saved_concept_ids=ids_to_delete)
-    print('Old topics deleted')
+    if len(ids_to_delete) > 0:
+        client.delete('concepts/saved', saved_concept_ids=ids_to_delete)
+        print('Old topics deleted')
+    else:
+        print('No Old topics to Delete')
         
 def copy_topics(client, topics):
     concepts = []
@@ -27,28 +30,25 @@ def main():
     parser.add_argument('from_url', help="The URL of the project to copy all topics from")
     parser.add_argument('to_url', help="The URL of the project to copy all topics into")
     parser.add_argument('--keep', default=False, action='store_true', help="Use this flag to specify if you want to keep the existing topics")
+    parser.add_argument('-t1', '--token1', default=None, help="Authentication token for first Daylight project")
+    parser.add_argument('-t2', '--token2', default=None, help="Authentication token for second Daylight project")
     
     args = parser.parse_args()
     from_proj = args.from_url.strip('/ ').split('/')[-1]
     from_root = args.from_url.split('/app')[0]
-    #from_root = '/'.join(args.from_url.strip('/ ').split('/')[:-4])
     to_proj = args.to_url.strip('/ ').split('/')[-1]
     to_root = args.to_url.split('/app')[0]
-    #to_root = '/'.join(args.to_url.strip('/ ').split('/')[:-4])
-    count = 0
-    while count < 3:
-        token = input('Token: ')
-        try:
-            from_client = LuminosoClient.connect(url='%s/api/v5/projects/%s' % (from_root, from_proj), token=token)
-            to_client = LuminosoClient.connect(url='%s/api/v5/projects/%s' % (to_root, to_proj), token=token)
-            break
-        except:
-            print('Incorrect credentials, please re-enter token')
-            count += 1
-            continue
-    if count >= 3:
-        print('Invalid credentials.')
-        return
+    if args.token1:
+        from_client = LuminosoClient.connect(url='%s/api/v5/projects/%s' % (from_root, from_proj), token=args.token1)
+    else:
+        from_client = LuminosoClient.connect(url='%s/api/v5/projects/%s' % (from_root, from_proj))
+    if args.token2:
+        to_client = LuminosoClient.connect(url='%s/api/v5/projects/%s' % (to_root, to_proj), token=args.token2)
+    elif args.token1 and not args.token2:
+        to_client = LuminosoClient.connect(url='%s/api/v5/projects/%s' % (to_root, to_proj), token=args.token1)
+    else:
+        to_client = LuminosoClient.connect(url='%s/api/v5/projects/%s' % (to_root, to_proj))
+        
     topics = from_client.get('concepts/saved')
     if not args.keep:
         delete_topics(to_client)
